requirements.txt
numpy==1.16.3
Pillow==6.0.0
pytorch-ignite==0.2.0
six==1.12.0
torch==1.0.1.post2
torchvision==0.2.2.post3
